requirements.txt
fastapi
uvicorn
supabase
python-dotenv
sqlalchemy
psycopg2-binary
psycopg2
pandas

